+++ build.sh
@@ -35,7 +35,7 @@ npm install
 grunt
 
 # Render SVG logos as PNG images
-inkscape src/svg/icon/48.svg --export-png=img/logo.png -w192 -h192
+inkscape src/svg/icon/48.svg --export-png=img/logo.png -w80 -h80
 inkscape src/svg/icon/square.svg --export-png=img/image.png --export-background=#ffffff  -w256 -h256
 
 inkscape src/svg/icon/square-precomposed.svg --export-png=homescreen-192x192.png -w192 -h192